requirements.txt
numpy==1.26.2
pandas==2.1.0
jinja2==3.1.2
pdfkit==1.0.0
Pillow==10.0.0
openpyxl==3.1.3